Move 256-bit vector definitions into their own module.
